src/backend/uac-registry/requirements.txt
# Django related
django==3.1.3
djangorestframework==3.12.2
django-extensions==3.1.0
django-cors-headers==3.5.0
djangorestframework-jwt==1.11.0

# Database related
psycopg2==2.8.6

# HTTP requests related
urllib3==1.26.2 

# Swagger OpenAPI related
drf-yasg[validation]==1.20.0
